+++ mobilecoind/src/test_utils.rs
@@ -182,6 +182,7 @@ pub fn setup_server<FPR: FogPubkeyResolver + Default + Send + Sync + 'static>(
         network_state,
         uri,
         None,
+        "unit-test".into(),
         logger,
     );